Services

+++ better/Models/City.cs
@@ -10,7 +10,7 @@ namespace better.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public double TransportCostToCity { get; set; }
-        public decimal PriceByHour { get; set; }
+        public double PriceByHour { get; set; }
         public virtual SearchHistory SearchHistory { get; set; }
 
     }
